+++ tools/swift-api-digester/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_swift_host_tool(swift-api-digester
   swift-api-digester.cpp
+  ModuleAnalyzerNodes.cpp
   LINK_LIBRARIES swiftFrontend swiftIDE
   SWIFT_COMPONENT tools
 )
